ai-service/requirements.txt
fastapi
uvicorn
pydantic
numpy
pandas
scikit-learn
mysql-connector-python
python-dotenv
